requirements.txt
aiodns==3.2.0 ; python_version >= "3.12" and python_version < "4.0"
aiohttp-cors==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.9.3 ; python_version >= "3.12" and python_version < "4.0"
aiokafka==0.10.0 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.3.0 ; python_version >= "3.12" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.12" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
ciso8601==2.3.1 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "win32" or platform_system == "Windows")
colorlog==6.8.2 ; python_version >= "3.12" and python_version < "4.0"
confluent-kafka==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
croniter==2.0.3 ; python_version >= "3.12" and python_version < "4.0"
cython==3.0.10 ; python_version >= "3.12" and python_version < "4.0"
dnspython==2.6.1 ; python_version >= "3.12" and python_version < "4.0"
email-validator==2.1.1 ; python_version >= "3.12" and python_version < "4.0"
fastapi[all]==0.110.1 ; python_version >= "3.12" and python_version < "4.0"
faust-cchardet==2.1.19 ; python_version >= "3.12" and python_version < "4.0"
faust-streaming-rocksdb==0.9.3 ; python_version >= "3.12" and python_version < "4.0"
faust-streaming[fast,prometheus,rocksdb,uvloop]==0.11.0 ; python_version >= "3.12" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.0.3 ; python_version >= "3.12" and python_version < "4.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.12" and python_version < "4.0"
httptools==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.6 ; python_version >= "3.12" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.12" and python_version < "4.0"
intervaltree==3.1.0 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.12" and python_version < "4.0"
jinja2==3.1.3 ; python_version >= "3.12" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.12" and python_version < "4.0"
jsonschema==4.21.1 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
mode-streaming==0.4.0 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.12" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
opentracing==2.4.0 ; python_version >= "3.12" and python_version < "4.0"
orjson==3.10.0 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.12" and python_version < "4.0"
pluggy==1.4.0 ; python_version >= "3.12" and python_version < "4.0"
prometheus-client==0.20.0 ; python_version >= "3.12" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.12" and python_version < "4.0"
pycares==4.4.0 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.16.3 ; python_version >= "3.12" and python_version < "4.0"
pydantic-extra-types==2.6.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic-settings==2.2.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.6.4 ; python_version >= "3.12" and python_version < "4.0"
pytest==8.1.1 ; python_version >= "3.12" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
python-multipart==0.0.9 ; python_version >= "3.12" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.12" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.12" and python_version < "4.0"
referencing==0.34.0 ; python_version >= "3.12" and python_version < "4.0"
rpds-py==0.18.0 ; python_version >= "3.12" and python_version < "4.0"
setproctitle==1.3.3 ; python_version >= "3.12" and python_version < "4.0"
setuptools==69.2.0 ; python_version >= "3.12" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sortedcontainers==2.4.0 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.29 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.37.2 ; python_version >= "3.12" and python_version < "4.0"
terminaltables==3.1.10 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.11.0 ; python_version >= "3.12" and python_version < "4.0"
ujson==5.9.0 ; python_version >= "3.12" and python_version < "4.0"
uvicorn[standard]==0.29.0 ; python_version >= "3.12" and python_version < "4.0"
uvloop==0.19.0 ; python_version >= "3.12" and python_version < "4.0"
venusian==3.1.0 ; python_version >= "3.12" and python_version < "4.0"
watchfiles==0.21.0 ; python_version >= "3.12" and python_version < "4.0"
websockets==12.0 ; python_version >= "3.12" and python_version < "4.0"
yarl==1.9.4 ; python_version >= "3.12" and python_version < "4.0"
